added motion framer to album.tracks
