+++ app/components/ui/albums.grid.tsx
@@ -4,7 +4,6 @@ import {
   Card,
   Image,
   Text,
-  Group,
   Badge,
   Title,
   Container,
